backend/requirements.txt
# Web & Infrastructure
flask
flask-cors
flask-socketio
eventlet
pyngrok
python-dotenv
requests
tqdm
pydantic

# core AI / ML
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
einops
numpy
scipy
opencv-python-headless
Pillow

# Face & Animation (LivePortrait / FaceSwap)
insightface
onnx
onnxruntime-gpu; platform_system == 'Linux'
onnxruntime; platform_system == 'Windows'
pyyaml
omegaconf
coloredlogs
gradio

# Media Processing (Audio/Video/Restoration)
edge-tts
openai-whisper

moviepy==1.0.3
imageio-ffmpeg
basicsr
gfpgan
realesrgan

# Tools
huggingface_hub
python-multipart
MarkupSafe
Werkzeug
jinja2
itsdangerous

PyJWT
Flask-Limiter
